Fix Bug related with the ids of the gradient's inputs
The bug was originated because the id of the inputs of the div gradients-options
was not a numeric value (0 and 1).
This is caused since the class Gradient has a method paint(e) that dynamically
associates the id of the color input to an array that stores the colors of the
current canvas element.